+++ Kernel/Security/Random/VirtIO/RNG.cpp
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-2-Clause
  */
 
-#include <Kernel/Bus/VirtIO/RNG.h>
+#include <Kernel/Security/Random/VirtIO/RNG.h>
 #include <Kernel/Bus/VirtIO/Transport/PCIe/TransportLink.h>
 #include <Kernel/Sections.h>
