+++ Meta/gn/secondary/Kernel/BUILD.gn
@@ -497,7 +497,6 @@ executable("Kernel_bin") {
     "Bus/USB/USBTransfer.cpp",
     "Bus/VirtIO/Device.cpp",
     "Bus/VirtIO/Queue.cpp",
-    "Bus/VirtIO/RNG.cpp",
     "Devices/AsyncDeviceRequest.cpp",
     "Devices/Audio/AC97/AC97.cpp",
     "Devices/Audio/Channel.cpp",
@@ -719,6 +718,7 @@ executable("Kernel_bin") {
     "Net/UDPSocket.cpp",
     "Net/VirtIO/VirtIONetworkAdapter.cpp",
     "SanCov.cpp",
+    "Security/Random/VirtIO/RNG.cpp",
     "Security/AddressSanitizer.cpp",
     "Security/Credentials.cpp",
     "Security/Jail.cpp",
